Separates classes into files
