--- modules/items.py
@@ -1,96 +1,0 @@
-from functools import reduce
-from modules.locatable import Locatable
-
-class Item(Locatable):
-    """
-    Item
-
-    This is the base class for items
-    """
-    def __init__(self, name=None, description=None):
-        Locatable.__init__(self)
-        self.name = name
-        self.description = description
-
-class Bag(Item):
-    """
-    Bag
-
-    This class is the incredible inventory bag.  Like the one from Mary Poppins, only cooler.
-    """
-    def __init__(self):
-        super().__init__(name="Bag of Holding", description="Like Mary Poppins' bag, this bag is infinitely large and made of tear-resistant duct tape.")
-        self.items = {}
-
-    def is_empty(self):
-        return self.item_count() == 0 
-
-    def item_count(self):
-        total = 0
-
-        for item in self.items.values():
-            total += item['count']
-
-        return total
-
-    def how_many(self, name):
-        if name in self.items:
-            return self.items[name]['count']
-        else:
-            return 0
-
-    def add(self, item):
-        self.add_many(item, 1)
-
-    def add_many(self, item, quantity):
-        if item.name in self.items:
-            self.items[item.name]['count'] += quantity
-        else:
-            self.items[item.name] = {'count': quantity, 'name':item.name, 'item': item}
-    
-    def remove(self, item_name):
-        return self.remove_many(item_name, 1)
-
-    def remove_many(self, item_name, quantity):
-        if item_name in self.items:
-            item_dict = self.items[item_name]
-
-            item = item_dict['item']
-            item_count = item_dict['count']
-            new_item_count = item_count - quantity
-
-            if item_count >= quantity:
-                item_dict['count'] -= quantity
-                taken = quantity
-            else:
-                taken = item_count
-
-            if new_item_count <= 0:
-                self.items.pop(item_name)
-
-            return (taken, item)
-        else:
-            return (0, None)
-
-    def dump(self):
-        return self.items
-
-    def format_item(self, item):
-        return ("{0} {1}\n".format(item['count'], item['name']))
-
-    def look(self):
-        seen = ""
-
-        for item in self.items:
-            seen = seen + self.format_item(self.items[item])
-
-        return seen
-
-class Weapon(Item):
-    """
-    This is the base class for a weapon. Each weapon instance takes
-    a name as a string and damage as an int.
-    """
-    def __init__(self, damage):
-        super().__init__()
-        self.damage = damage
