+++ modules/locatable.py
@@ -1,6 +1,11 @@
+# locatable.py
+""" In support of the Level class, Locatable module contains code that is
+location aware
+"""
 from uuid import uuid4
 
 class Locatable:
+    """The ancestor class for stuff in a Level that is aware of where it is"""
     def __init__(self):
         self.uid = uuid4()
         self.coords = (0, 0)
@@ -9,18 +14,26 @@ class Locatable:
         self.move_ai = False
 
     def place(self, coords):
+        """Sets the coordinates of the object"""
         self.coords = coords
 
     def set_display(self, display):
+        """Returns the symbol that represented this object on the map"""
         self.display = display
 
     def is_at(self, coords):
+        """Used to determine if this object is located as specific
+        coordinates
+        """
         return self.coords == coords
 
     def locate(self):
+        """Returns the object's current location"""
         return self.coords
 
-    def go(self, direction):
+    def travel(self, direction):
+        """Given a character representing a cardinal direction
+        this method will move the object in that direction"""
         if direction == "n":
             self.coords = (self.coords[0], self.coords[1] + 1)
         elif direction == "s":
@@ -31,4 +44,5 @@ class Locatable:
             self.coords = (self.coords[0] - 1, self.coords[1])
 
     def has_move_ai(self):
+        """Used to indicate that this object will want to move"""
         return self.move_ai
