+++ pylint.output
@@ -1,0 +1,125 @@
+************* Module modules.level
+R: 95, 4: Method could be a function (no-self-use)
+R:109, 4: Method could be a function (no-self-use)
+W:123, 8: Attempting to unpack a non-sequence defined at line 43 (unpacking-non-sequence)
+************* Module modules.engine
+R: 15, 0: Too many instance attributes (9/7) (too-many-instance-attributes)
+
+
+Report
+======
+459 statements analysed.
+
+Statistics by type
+------------------
+
++---------+-------+-----------+-----------+------------+---------+
+|type     |number |old number |difference |%documented |%badname |
++=========+=======+===========+===========+============+=========+
+|module   |14     |14         |=          |100.00      |0.00     |
++---------+-------+-----------+-----------+------------+---------+
+|class    |10     |10         |=          |100.00      |0.00     |
++---------+-------+-----------+-----------+------------+---------+
+|method   |70     |70         |=          |100.00      |0.00     |
++---------+-------+-----------+-----------+------------+---------+
+|function |9      |9          |=          |100.00      |0.00     |
++---------+-------+-----------+-----------+------------+---------+
+
+
+
+External dependencies
+---------------------
+::
+
+    modules 
+      \-bag (modules.engine)
+      \-command_line (modules.engine)
+      \-level (modules.level_loader)
+      \-level_loader (modules.engine)
+      \-locatable (modules.item,modules.player,modules.monsters)
+      \-monsters (modules.level_loader)
+      \-movement (modules.monsters)
+      \-player (modules.engine)
+      \-world (modules.engine)
+
+
+
+Raw metrics
+-----------
+
++----------+-------+------+---------+-----------+
+|type      |number |%     |previous |difference |
++==========+=======+======+=========+===========+
+|code      |548    |73.26 |548      |=          |
++----------+-------+------+---------+-----------+
+|docstring |144    |19.25 |144      |=          |
++----------+-------+------+---------+-----------+
+|comment   |13     |1.74  |13       |=          |
++----------+-------+------+---------+-----------+
+|empty     |43     |5.75  |43       |=          |
++----------+-------+------+---------+-----------+
+
+
+
+Duplication
+-----------
+
++-------------------------+------+---------+-----------+
+|                         |now   |previous |difference |
++=========================+======+=========+===========+
+|nb duplicated lines      |0     |0        |=          |
++-------------------------+------+---------+-----------+
+|percent duplicated lines |0.000 |0.000    |=          |
++-------------------------+------+---------+-----------+
+
+
+
+Messages by category
+--------------------
+
++-----------+-------+---------+-----------+
+|type       |number |previous |difference |
++===========+=======+=========+===========+
+|convention |0      |0        |=          |
++-----------+-------+---------+-----------+
+|refactor   |3      |3        |=          |
++-----------+-------+---------+-----------+
+|warning    |1      |1        |=          |
++-----------+-------+---------+-----------+
+|error      |0      |0        |=          |
++-----------+-------+---------+-----------+
+
+
+
+% errors / warnings by module
+-----------------------------
+
++---------------+------+--------+---------+-----------+
+|module         |error |warning |refactor |convention |
++===============+======+========+=========+===========+
+|modules.level  |0.00  |100.00  |66.67    |0.00       |
++---------------+------+--------+---------+-----------+
+|modules.engine |0.00  |0.00    |33.33    |0.00       |
++---------------+------+--------+---------+-----------+
+
+
+
+Messages
+--------
+
++-----------------------------+------------+
+|message id                   |occurrences |
++=============================+============+
+|no-self-use                  |2           |
++-----------------------------+------------+
+|unpacking-non-sequence       |1           |
++-----------------------------+------------+
+|too-many-instance-attributes |1           |
++-----------------------------+------------+
+
+
+
+Global evaluation
+-----------------
+Your code has been rated at 9.91/10 (previous run: 9.91/10, +0.00)
+
