+++ tests/test_user_interface.py
@@ -1,5 +1,5 @@
 import pytest
-from modules.world import CommandLineInterface
+from modules.command_line import CommandLine
 from os import getcwd
 from os.path import join
 
@@ -9,6 +9,6 @@ class TestKnot:
     @pytest.fixture()
     def prepare_ui(self):
         pass
- 
+
     def test_knot_will_accept_story_path(self):
         pass
